+++ src/Campaign.php
@@ -180,7 +180,7 @@ class Campaign extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '3.0.0.3';
+    public string $schemaVersion = '3.1.2';
 
     /**
      * @inheritdoc